% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eifs.R
\name{compute_eif}
\alias{compute_eif}
\title{Efficient Influence Function Computer}
\usage{
compute_eif(
  data,
  effect,
  confounders,
  exposure,
  outcome,
  modifiers,
  prop_score_fit,
  prop_score_values,
  cond_outcome_fit,
  failure_hazard_fit,
  censoring_hazard_fit,
  ace_estimate,
  plugin_estimates
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data.
\code{train_data} is formatted by \code{\link{unihtee}()}.}

\item{effect}{A \code{character} indicating the type of treatment effect
modifier variable importance parameter. Currently supports
\code{"absolute"} and \code{"relative"}.}

\item{confounders}{A \code{character} vector of column names corresponding to
baseline covariates.}

\item{exposure}{A \code{character} corresponding to the exposure variable.}

\item{outcome}{A \code{character} corresponding to the outcome variable.}

\item{modifiers}{A \code{character} vector of columns names corresponding to
the suspected effect modifiers. This vector must be a subset of
\code{confounders}.}

\item{prop_score_fit}{A \code{list} output by the
\code{\link{fit_prop_score}()} function.}

\item{prop_score_values}{A \code{numeric} vector corresponding to the (known)
propensity score values for each observation in \code{data}.}

\item{cond_outcome_fit}{A \code{list} output by the
\code{\link{fit_failure_hazard}()} function.}

\item{failure_hazard_fit}{A \code{list} output by the
\code{\link{fit_cond_outcome}()} function.}

\item{censoring_hazard_fit}{A \code{list} output by the
\code{\link{fit_censoring_hazard}()} function.}

\item{ace_estimate}{A \code{numeric} estimate of the average causal effect
associated with the specified effect type and data-generating process.}

\item{plugin_estimates}{A \code{numeric} vector of plug-in estimate of the
treatment effect modifier variable importance parameters associated with
the specified effect type and data-generating process.}
}
\value{
A \code{data.table} whose columns are the efficient influence
functions of each variable in \code{modifiers}. The rows correspond to the
observations of \code{data}.
}
\description{
\code{compute_eif()} computes the efficient influence function
for the chosen parameter using the already estimated nuisance parameters.
If certain nuisance parameters are known, such as propensity scores in a
randomized control trial, then they may be input directly into this
function. Note that this EIF is only uncentered in expectation; it should
only be used with the one-step estimator. When \code{ace_estimate} is
\code{NULL}, and pseudo-uncentered EIF is computed. It should only be used
with the one-step estimator, as it is only uncentered in expectation.
}
\keyword{internal}
