% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance-estimators.R
\name{fit_cond_outcome}
\alias{fit_cond_outcome}
\title{Conditional Outcome Estimator}
\usage{
fit_cond_outcome(
  train_data,
  valid_data,
  learners,
  outcome,
  exposure,
  confounders
)
}
\arguments{
\item{train_data}{A \code{data.table} containing the observed data.
\code{train_data} is formatted by \code{\link{unihtee}()}.}

\item{valid_data}{An optional \code{data.table} representing a holdout
dataset of the observed data. It is only used for cross-fitting purposes.
Defaults to \code{NULL}.}

\item{learners}{A \code{\link[sl3]{Stack}}, or other learner class
(inheriting from \code{\link[sl3]{Lrnr_base}}), containing a set of
learners from \pkg{sl3} to estimate the propensity score model.}

\item{outcome}{A \code{character} corresponding to the outcome variable.}

\item{exposure}{A \code{character} corresponding to the exposure variable.}

\item{confounders}{A \code{character} vector of column names corresponding to
baseline covariates.}
}
\value{
A named \code{list} of two elements. (1) \code{"estimates"}, the
expected conditional outcome for each observation in \code{valid_data}, if
specified, or \code{train_data} otherwise. (2) \code{"fit"}, the trained
\code{\link[sl3]{Stack}} or learner.
}
\description{
\code{fit_cond_outcome()} estimates the conditional outcome,
based on the exposure and confounders, over the \code{train_data} dataset.
The estimator used for this estimation is based on the \code{learners}
argument, and the covariates considered are specified by the
\code{confounders} and \code{exposure} arguments.
}
\keyword{internal}
