% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance-estimators.R
\name{fit_prop_score}
\alias{fit_prop_score}
\title{Propensity Score Estimator}
\usage{
fit_prop_score(train_data, valid_data, learners, exposure, confounders)
}
\arguments{
\item{train_data}{A \code{data.table} containing the observed data.
\code{train_data} is formatted by \code{\link{unihtee}()}.}

\item{valid_data}{An optional \code{data.table} representing a holdout
dataset of the observed data. It is only used for cross-fitting purposes.
Defaults to \code{NULL}.}

\item{learners}{A \code{\link[sl3]{Stack}}, or other learner class
(inheriting from \code{\link[sl3]{Lrnr_base}}), containing a set of
learners from \pkg{sl3} to estimate the propensity score model.}

\item{exposure}{A \code{character} corresponding to the exposure variable.}

\item{confounders}{A \code{character} vector of column names corresponding to
baseline covariates.}
}
\value{
A named \code{list} of two elements. (1) \code{"estimates"}, the
propensity score estimates for each observation in \code{valid_data}, if
specified, or \code{train_data} otherwise. (2) \code{"fit"}, the trained
\code{\link[sl3]{Stack}} or learner.
}
\description{
\code{fit_prop_score()} estimates the propensity score over the
\code{train_data} dataset. The estimator used for this estimation is based
on the \code{learners} argument, and the covariates considered are
specified by the \code{confounders} argument.
}
\keyword{internal}
