% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{one_step_estimator}
\alias{one_step_estimator}
\title{One Step Estimator}
\usage{
one_step_estimator(uncentered_eif_data)
}
\arguments{
\item{uncentered_eif_data}{A \code{data.table} where each column is an
efficient influence function of a potential treatment effect modifier
applied to some dataset.}
}
\value{
A one-row \code{data.table} containing the one-step estimates for
each potential modifier.
}
\description{
\code{one_step_estimator()} computes the one-step estimates of
any given parameter when provided with uncentered efficient influence
functions.
}
\keyword{internal}
