% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_temvip}
\alias{plot_temvip}
\title{Plot Treatment Effect Modifier Variable Importance Parameter Estimate}
\usage{
plot_temvip(unihtee_output, modifier_name, print_interpretation = TRUE)
}
\arguments{
\item{unihtee_output}{A \code{unihtee} class object output by
\code{\link{unihtee}()}.}

\item{print_interpretation}{A \cpde{flag} indicating whether to print the
interpretation of the TEM-VIP inference plot. Defaults to \code{TRUE}.}

\item{modifier}{A \code{character} specifying the treatment effect modifier
variable importance parameter estimate to plot.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object.
}
\description{
\code{plot.unihtee()} produces a plot depicting the estimated
simple linear regression line for the specified \code{modifier} parameter
estimated by \code{\link{unihtee}()}.
}
