% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.unihtee}
\alias{print.unihtee}
\title{Print a \code{unihtee} Object}
\usage{
\method{print}{unihtee}(x, ...)
}
\arguments{
\item{x}{A \code{unihtee} object.}

\item{...}{Ignored.}
}
\description{
Print a \code{unihtee} Object
}
