% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance-estimators.R
\name{tml_ate_estimator}
\alias{tml_ate_estimator}
\title{Targeted Maximum Likelihood Estimator of the Average Treatment Effect}
\usage{
tml_ate_estimator(
  data,
  confounders,
  exposure,
  outcome,
  prop_score_fit,
  prop_score_values = NULL,
  cond_outcome_fit
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data.
\code{train_data} is formatted by \code{\link{unihtee}()}.}

\item{confounders}{A \code{character} vector of column names corresponding to
baseline covariates.}

\item{exposure}{A \code{character} corresponding to the exposure variable.}

\item{outcome}{A \code{character} corresponding to the outcome variable.}

\item{prop_score_fit}{A \code{list} output by the
\code{\link{fit_prop_score}()} function.}

\item{prop_score_values}{A \code{numeric} vector corresponding to the (known)
propensity score values for each observation in \code{data}.}

\item{cond_outcome_fit}{A \code{list} output by the
\code{\link{fit_cond_outcome}()} function.}
}
\value{
The targeted maximum likelihood estimate of the average treatment
effect.
}
\description{
\code{tml_ate_estimator()} implements a targeted maximum likelihood estimator of
the average treatment effect.
}
\keyword{internal}
