% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{tml_estimator}
\alias{tml_estimator}
\title{Targeted Maximum Likelihood Estimator}
\usage{
tml_estimator(
  data,
  confounders,
  modifiers,
  exposure,
  outcome,
  scale,
  prop_score_fit,
  prop_score_values = NULL,
  cond_outcome_fit,
  failure_hazard_fit,
  censoring_hazard_fit
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data.
\code{train_data} is formatted by \code{\link{unihtee}()}.}

\item{confounders}{A \code{character} vector of column names corresponding to
baseline covariates.}

\item{modifiers}{A \code{character} vector of columns names corresponding to
the suspected effect modifiers. This vector must be a subset of
\code{confounders}.}

\item{exposure}{A \code{character} corresponding to the exposure variable.}

\item{outcome}{A \code{character} corresponding to the outcome variable.}

\item{scale}{A \code{character} indicating the type of treatment effect
modifier variable importance parameter. Currently supports
\code{"absolute"} and \code{"relative"}.}

\item{prop_score_fit}{A \code{list} output by the
\code{\link{fit_prop_score}()} function.}

\item{prop_score_values}{A \code{numeric} vector corresponding to the (known)
propensity score values for each observation in \code{data}.}

\item{cond_outcome_fit}{A \code{list} output by the
\code{\link{fit_cond_outcome}()} function.}

\item{failure_hazard_fit}{A \code{list} output by the
\code{\link{fit_cond_outcome}()} function.'}

\item{censoring_hazard_fit}{A \code{list} output by the
\code{\link{fit_censoring_hazard}()} function.'}
}
\value{
A one-row \code{data.table} containing the targeted maximum
likelihood estimates for each potential modifier.
}
\description{
\code{tml_estimator()} computes the targeted maximum likelihood
estimates of any given parameter.
}
\keyword{internal}
