% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance-estimators.R
\name{tml_estimator_ate_log_outcome}
\alias{tml_estimator_ate_log_outcome}
\title{TMLE of the Marginal Relative Effect for Non-Time-to-Event Outcomes}
\usage{
tml_estimator_ate_log_outcome(
  data,
  confounders,
  exposure,
  outcome,
  prop_score_fit,
  prop_score_values = NULL,
  cond_outcome_fit
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data.
\code{train_data} is formatted by \code{\link{unihtee}()}.}

\item{confounders}{A \code{character} vector of column names corresponding to
baseline covariates.}

\item{exposure}{A \code{character} corresponding to the exposure variable.}

\item{outcome}{A \code{character} corresponding to the outcome variable.}

\item{prop_score_fit}{A \code{list} output by the
\code{\link{fit_prop_score}()} function.}

\item{prop_score_values}{A \code{numeric} vector corresponding to the (known)
propensity score values for each observation in \code{data}.}

\item{cond_outcome_fit}{A \code{list} output by the
\code{\link{fit_cond_outcome}()} function.}
}
\value{
The TML estimate.
}
\description{
TMLE of the Marginal Relative Effect for Non-Time-to-Event Outcomes
}
\keyword{internal}
