% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance-estimators.R
\name{tml_rmst_diff_estimator}
\alias{tml_rmst_diff_estimator}
\title{Targeted Maximum Likelihood Estimator of Difference of RMST}
\usage{
tml_rmst_diff_estimator(
  data,
  exposure,
  outcome,
  prop_score_fit,
  prop_score_values,
  failure_hazard_fit,
  censoring_hazard_fit
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data.
\code{train_data} is formatted by \code{\link{unihtee}()}.}

\item{exposure}{A \code{character} corresponding to the exposure variable.}

\item{outcome}{A \code{character} corresponding to the outcome variable.}

\item{prop_score_fit}{A \code{list} output by the
\code{\link{fit_prop_score}()} function.}

\item{prop_score_values}{A \code{numeric} vector corresponding to the (known)
propensity score values for each observation in \code{data}.}

\item{failure_hazard_fit}{A \code{list} output by the
\code{\link{fit_cond_outcome}()} function.'}

\item{censoring_hazard_fit}{A \code{list} output by the
\code{\link{fit_censoring_hazard}()} function.'}
}
\value{
A \code{numeric} one-step estimate.
}
\description{
Targeted Maximum Likelihood Estimator of Difference of RMST
}
\keyword{internal}
