% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_data_melt.R
\name{tte_data_melt}
\alias{tte_data_melt}
\title{Time-to-Event Data Melt}
\usage{
tte_data_melt(
  data,
  confounders,
  exposure,
  outcome,
  censoring,
  time_cutoff,
  prop_score_values
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data.}

\item{confounders}{A \code{character} vector of column names corresponding to
baseline covariates.}

\item{exposure}{A \code{character} corresponding to the exposure variable.}

\item{outcome}{A \code{character} corresponding to the outcome variable.}

\item{censoring}{A \code{character} indicating the right censoring indicator
variable.}

\item{time_cutoff}{A \code{numeric} representing the time point at which to
evaluate the time-to-event parameter.}

\item{prop_score_values}{A \code{character} corresponding to the (known)
propensity score values for each observation in \code{data}.}
}
\value{
A long, time-to-event version of the original \code{data.table}
object. That is, observations are repeated at time points between the
earliest time in the data and the \code{time_cutoff} argument.
}
\description{
\code{tte_data_melt()} turns wide-format \code{data.table}
objects into long-format \code{data.table} objects.
}
\keyword{internal}
