% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eifs.R
\name{uncentered_eif}
\alias{uncentered_eif}
\title{Uncentered Efficient Influence Function Computer}
\usage{
uncentered_eif(
  data,
  confounders,
  exposure,
  outcome,
  modifiers,
  prop_score_fit,
  prop_score_values,
  cond_outcome_fit
)
}
\value{
A \code{data.table} whose columns are the uncentered efficient
influence functions of each variable in \code{modifiers}. The rows
correspond to the observations of \code{data}.
}
\description{
\code{uncentered_eif()} computes thei efficient influence
function for the chosen parameter using the already estimated nuisance
parameters. If certain nuisance parameters are known, such as propensity
scores in a randomized control trial, then they may be input directly into
this function.
}
\keyword{internal}
