% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.unihtee}
\alias{plot.unihtee}
\title{Plot a \code{unihtee} Object}
\usage{
\method{plot}{unihtee}(x, y, ..., modifier_name, print_interpretation = TRUE)
}
\arguments{
\item{x}{A \code{unihtee} class object output by \code{\link{unihtee}()}.}

\item{y}{Ignored.}

\item{...}{Ignored.}

\item{modifier_name}{A \code{character} specifying the treatment effect
modifier variable importance parameter estimate to plot.}

\item{print_interpretation}{A \code{flag} indicating whether to print the
interpretation of the TEM-VIP inference plot. Defaults to \code{TRUE}.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object.
}
\description{
\code{plot.unihtee()} produces a plot depicting the estimated
simple linear regression line for the specified \code{modifier} parameter
estimated by \code{\link{unihtee}()}.
}
