% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unihtee.R
\name{unihtee}
\alias{unihtee}
\title{Univariate Heterogeneous Treatment Effect Modifier Estimator}
\usage{
unihtee(
  data,
  confounders,
  modifiers,
  exposure,
  outcome,
  censoring = NULL,
  time_cutoff = NULL,
  outcome_type = c("continuous", "binary", "time-to-event"),
  effect = c("absolute", "relative"),
  estimator = c("tmle", "onestep"),
  cross_fit = FALSE,
  cross_fit_folds = 5,
  cond_outcome_estimator = sl3::Lrnr_glm_fast$new(),
  prop_score_estimator = sl3::Lrnr_glm_fast$new(),
  prop_score_values = NULL,
  failure_hazard_estimator = sl3::Lrnr_xgboost$new(),
  censoring_hazard_estimator = sl3::Lrnr_xgboost$new(),
  parallel = FALSE
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data.}

\item{confounders}{A \code{character} vector of column names corresponding to
baseline covariates.}

\item{modifiers}{A \code{character} vector of columns names corresponding to
the suspected effect modifiers. This vector must be a subset of
\code{confounders}.}

\item{exposure}{A \code{character} corresponding to the exposure variable.}

\item{outcome}{A \code{character} corresponding to the outcome variable.}

\item{censoring}{A \code{character} indicating the right censoring indicator
variable. Only used with time-to-event outcomes. Defaults to \code{NULL}.}

\item{time_cutoff}{A \code{numeric} representing the time point at which to
evaluate the time-to-event parameter. Only used with time-to-event
outcomes. Defaults to \code{NULL}.}

\item{outcome_type}{A \code{character} indicating the outcome type.
\code{"continuous"}, \code{"binary"} and \code{"time-to-event"} are
currently supported.}

\item{effect}{A \code{character} indicating the type of treatment effect
modifier variable importance parameter. Currently supports
\code{"absolute"} and \code{"relative"}.}

\item{estimator}{A \code{character} set to either \code{"tmle"} or
\code{"onestep"}. The former results in \code{unihtee()} to use a targeted
maximum likelihood estimators to estimate the desired TEM-VIP, while the
latter uses a one step estimator.}

\item{cross_fit}{A \code{logical} determining whether cross-fitting should be
used. Defaults to \code{FALSE}.}

\item{cross_fit_folds}{A \code{numeric} stating the number of folds to use in
the cross-fitting procedure. Defaults to 5.}

\item{cond_outcome_estimator}{A \code{\link[sl3]{Stack}}, or other learner
class (inheriting from \code{\link[sl3]{Lrnr_base}}), containing a set of
learners from \pkg{sl3} to estimate the conditional outcome. Defaults to a
generalized linear model with one- and two- way interactions among all
\code{confounders} and \code{exposure} variables. Only used with continuous
and binary outcomes.}

\item{prop_score_estimator}{A \code{\link[sl3]{Stack}}, or other learner
class (inheriting from \code{\link[sl3]{Lrnr_base}}), containing a set of
learners from \pkg{sl3} to estimate the propensity score. Defaults to a
generalized linear model with one- and two- way interactions among all
\code{confounders} variables.}

\item{prop_score_values}{An optional \code{character} corresponding to the
(known) propensity score values for each observation in \code{data}.
Defaults to \code{NULL}.}

\item{failure_hazard_estimator}{A \code{\link[sl3]{Stack}}, or other learner
class (inheriting from \code{\link[sl3]{Lrnr_base}}), containing a set of
learners from \pkg{sl3} to estimate the conditional failure hazard
function. Defaults to an XGBoost learner with \code{confounders} and
\code{exposure} variables as covariates. Only used with time-to-event
outcomes.}

\item{censoring_hazard_estimator}{A \code{\link[sl3]{Stack}}, or other
learner class (inheriting from \code{\link[sl3]{Lrnr_base}}), containing a
set of learners from \pkg{sl3} to estimate the conditional censoring hazard
function. Defaults to an XGBoost learner with \code{confounders} and
\code{exposure} variables as covariates. Only used with time-to-event
outcomes.}

\item{parallel}{A \code{logical} stating if
\code{\link[origami:cross_validate]{origami}}'s built-in parallelized
cross-validation routines should be used when \code{cross_fit = TRUE}. The
\href{https://cran.r-project.org/package=future}{\code{future}} suite is
used. Defaults to \code{FALSE}.}
}
\value{
A list containing:
\itemize{
\item \code{temvip_inference_tbl}: A \code{data.table} containing the effect
estimates and (adjusted) p-values of the \code{modifiers}. The suspected
treatment effect modifiers ordered according to ascending p-values.
\item \code{ace_estimate}: A \code{numeric} providing the estimate of the
average causal effect associated with the specified effect and outcome
types.
\item \code{data}: The \code{data.table} containing the observed data used to
estimate the TEM-VIPs, containing only the confounders, modifiers,
exposure, outcome, censoring (if provided), and propensity score
(if provided) variables.
}
}
\description{
\code{unihtee()} estimates treatment effect modifier variable
importance parameters (TEM-VIPs). Both absolute and relative TEM-VIPs can
be estimated using one-step or targeted maximum likelihood estimators.
}
