% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unihtee.R
\name{unihtee}
\alias{unihtee}
\title{Univariate Heterogeneous Treatment Effect Modifier Estimator}
\usage{
unihtee(
  data,
  confounders,
  modifiers,
  exposure,
  outcome,
  outcome_type = c("continuous", "binary"),
  estimator = c("tmle", "onestep"),
  cond_outcome_estimator = sl3::Lrnr_glm_fast$new(),
  prop_score_estimator = sl3::Lrnr_glm_fast$new(),
  prop_score_values = NULL
)
}
\arguments{
\item{data}{A \code{data.table} containing the observed data.}

\item{confounders}{A \code{character} vector of column names corresponding to
baseline covariates.}

\item{modifiers}{A \code{character} vector of columns names corresponding to
the suspected effect modifiers. This vector must be a subset of
\code{confounders}.}

\item{exposure}{A \code{character} corresponding to the exposure variable.}

\item{outcome}{A \code{character} corresponding to the outcome variable.}

\item{outcome_type}{A \code{character} indicating the outcome type.
\code{"continuous"} and \code{"binary"} are currently supported.}

\item{estimator}{A \code{character} set to either \code{"tmle"} or
\code{"onestep"}. The former results in \code{unihtee()} to use a targeted
maximum likelihood estimators to estimate the deisred TEM VIP, while the
latter uses a one step estimator.}

\item{cond_outcome_estimator}{A \code{\link[sl3]{Stack}}, or other learner
class (inheriting from \code{\link[sl3]{Lrnr_base}}), containing a set of
learners from \pkg{sl3} to estimate the propensity score model. Defaults to
a generalized linear model with one- and two- way interactions among all
\code{confounders} and \code{exposure} variables.}

\item{prop_score_estimator}{A \code{\link[sl3]{Stack}}, or other learner
class (inheriting from \code{\link[sl3]{Lrnr_base}}), containing a set of
learners from \pkg{sl3} to estimate the propensity score model. Defaults to
a generalized linear model with one- and two- way interactions among all
\code{confounders} variables.}

\item{prop_score_values}{A \code{numeric} vector corresponding to the (known)
propensity score values for each observation in \code{data}.}
}
\value{
A \code{data.table} containing the effect estimates and (adjusted)
p-values of the \code{modifiers}. The suspected treatment effect modifiers
ordered according to ascending p-values.
}
\description{
\code{unihtee()} estimates treatment effect modifiers variable
importance parameters (TEM VIPs). These TEM VIPs are defined on the risk
difference or relative risk scales and can be estimated using one step or
targeted maximum likelihood estimators.
}
